Restrict Access by Mobile
This plugin would allow us to restrict activities based on whether they are using their phones or if they are on their desktop. This can be useful for Nadia and Kate if they went ahead with Poodll or H5P as a temp fix until it's compatible.

Then we would have to worry about it being compatible because mobile users won't see it. And it's also supplemental so students won't be missing the main content.